+++ bonus/src/parsing/assignment_amb_bonus.c
@@ -58,7 +58,7 @@ void	parse_ambient(char **tokens, t_minirt *rt)
 	if (count_tokens(tokens) != 3)
 		ft_error("wrong elements");
 	rt->ambient.ratio = ft_atof(tokens[1]);
-	if (rt->ambient.ratio < 0.0 || rt->ambient.ratio > 1.0)
+	if (rt->ambient.ratio < 0.0f || rt->ambient.ratio > 1.0f)
 		ft_error("Ambient out of range");
 	rt->ambient.color = parse_color(tokens[2]);
 }
@@ -66,7 +66,7 @@ void	parse_ambient(char **tokens, t_minirt *rt)
 void	parse_light(char **tokens, t_minirt *rt)
 {
 	t_vector	position;
-	double		ratio;
+	float		ratio;
 	t_color		color;
 	t_light		*new_light;
 
@@ -79,7 +79,7 @@ void	parse_light(char **tokens, t_minirt *rt)
 		ft_error("wrong format of light");
 	position = parse_vector(tokens[1]);
 	ratio = ft_atof(tokens[2]);
-	if (ratio < 0.0 || ratio > 1.0)
+	if (ratio < 0.0f || ratio > 1.0f)
 		ft_error("Light ratio out of range");
 	new_light = light_lstnew(position, ratio, color);
 	light_lstadd_back(&rt->light, new_light);
