+++ src/math_tool/color.c
@@ -6,26 +6,26 @@
 /*   By: ymiao <ymiao@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/14 17:54:58 by ymiao             #+#    #+#             */
-/*   Updated: 2025/09/02 03:30:16 by ymiao            ###   ########.fr       */
+/*   Updated: 2025/09/05 01:06:53 by ymiao            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "math_tool.h"
 
 /**
- * @brief make int to double to get more accurencies
+ * @brief make int to float to get more accurencies
  */
 t_color	color_init(int r, int g, int b)
 {
 	t_color	res;
 
-	res.r = (double)r / 255.0;
-	res.g = (double)g / 255.0;
-	res.b = (double)b / 255.0;
+	res.r = (float)r / 255.0f;
+	res.g = (float)g / 255.0f;
+	res.b = (float)b / 255.0f;
 	return (res);
 }
 
-t_color	color_init_d(double r, double g, double b)
+t_color	color_init_d(float r, float g, float b)
 {
 	t_color	res;
 
@@ -45,7 +45,7 @@ t_color	color_add(t_color a, t_color b)
 	return (res);
 }
 
-t_color	color_div(t_color a, double scalar)
+t_color	color_div(t_color a, float scalar)
 {
 	t_color	res;
 
@@ -61,9 +61,9 @@ int	color_to_int(t_color color)
 	int	g;
 	int	b;
 
-	r = (int)(fmin(color.r, 1.0) * 255.0);
-	g = (int)(fmin(color.g, 1.0) * 255.0);
-	b = (int)(fmin(color.b, 1.0) * 255.0);
+	r = (int)(fminf(color.r, 1.0f) * 255.0f);
+	g = (int)(fminf(color.g, 1.0f) * 255.0f);
+	b = (int)(fminf(color.b, 1.0f) * 255.0f);
 	if (r < 0)
 		r = 0;
 	if (g < 0)
