+++ src/render/all_lights.c
@@ -42,14 +42,14 @@ static t_color	calculate_diffuse_light(t_light light, t_vector hit_point,
 								t_vector normal, t_color obj_color)
 {
 	t_vector	light_dir;
-	double		light_intensity;
+	float		light_intensity;
 	t_color		result;
 
 	light_dir = vector_sub(light.position, hit_point);
 	light_dir = vector_normalize(light_dir);
 	light_intensity = vector_dot(normal, light_dir);
-	if (light_intensity < 0.0)
-		light_intensity = 0.0;
+	if (light_intensity < 0.0f)
+		light_intensity = 0.0f;
 	light_intensity *= light.ratio;
 	result.r = light_intensity * obj_color.r;
 	result.g = light_intensity * obj_color.g;
@@ -82,11 +82,11 @@ t_color	combine_light(t_minirt *rt,
 	final_color.r = ambient_color.r + diffuse_color.r;
 	final_color.g = ambient_color.g + diffuse_color.g;
 	final_color.b = ambient_color.b + diffuse_color.b;
-	if (final_color.r > 1.0)
-		final_color.r = 1.0;
-	if (final_color.g > 1.0)
-		final_color.g = 1.0;
-	if (final_color.b > 1.0)
-		final_color.b = 1.0;
+	if (final_color.r > 1.0f)
+		final_color.r = 1.0f;
+	if (final_color.g > 1.0f)
+		final_color.g = 1.0f;
+	if (final_color.b > 1.0f)
+		final_color.b = 1.0f;
 	return (final_color);
 }
